more c# templatization; fix ai dev new list test
