+++ ideas/template-engine/refactor.yaml
@@ -16,6 +16,12 @@
     # - { "dir1": "openai-asst-cs", "dir2": "openai-asst-streaming-with-functions-cs",   "file1": "OpenAIAssistantsClass.cs",  "file2": "OpenAIAssistantsFunctionsStreamingClass.cs",        "file3": "OpenAIAssistantsClass.cs",  "dir3": "openai-asst-cs", "flag1": "!{_IS_OPENAI_ASST_STREAMING_TEMPLATE}", "flag2": "{_IS_OPENAI_ASST_FUNCTIONS_TEMPLATE}" }
     # - { "dir1": "openai-asst-cs", "dir2": "openai-asst-streaming-with-functions-cs",   "file1": "Program.cs",                "file2": "Program.cs",                                        "file3": "Program.cs",                "dir3": "openai-asst-cs", "flag1": "!{_IS_OPENAI_ASST_STREAMING_TEMPLATE}", "flag2": "{_IS_OPENAI_ASST_FUNCTIONS_TEMPLATE}" }
 
+    # Phi3 Chat
+    - { "dir1": "phi3-onnx-chat-streaming-cs", "dir2": "phi3-onnx-chat-streaming-with-functions-cs", "file1": "get-phi3-mini-onnx.cmd",                    "file2": "get-phi3-mini-onnx.cmd",                    "file3": "get-phi3-mini-onnx.cmd",                    "dir3": "phi3-onnx-chat-cs", "flag1": "{_IS_PHI3_ONNX_CHAT_STREAMING_TEMPLATE}", "flag2": "{_IS_PHI3_ONNX_CHAT_FUNCTIONS_TEMPLATE}" }
+    - { "dir1": "phi3-onnx-chat-streaming-cs", "dir2": "phi3-onnx-chat-streaming-with-functions-cs", "file1": "OnnxGenAIChatCompletionsStreamingClass.cs", "file2": "OnnxGenAIChatCompletionsStreamingClass.cs", "file3": "OnnxGenAIChatCompletionsStreamingClass.cs", "dir3": "phi3-onnx-chat-cs", "flag1": "{_IS_PHI3_ONNX_CHAT_STREAMING_TEMPLATE}", "flag2": "{_IS_PHI3_ONNX_CHAT_FUNCTIONS_TEMPLATE}" }
+    - { "dir1": "phi3-onnx-chat-streaming-cs", "dir2": "phi3-onnx-chat-streaming-with-functions-cs", "file1": "Phi3ChatStreaming.csproj._",                "file2": "Phi3ChatStreaming.csproj._",                "file3": "Phi3ChatStreaming.csproj._",                "dir3": "phi3-onnx-chat-cs", "flag1": "{_IS_PHI3_ONNX_CHAT_STREAMING_TEMPLATE}", "flag2": "{_IS_PHI3_ONNX_CHAT_FUNCTIONS_TEMPLATE}" }
+    - { "dir1": "phi3-onnx-chat-streaming-cs", "dir2": "phi3-onnx-chat-streaming-with-functions-cs", "file1": "Program.cs",                                "file2": "Program.cs",                                "file3": "Program.cs",                                "dir3": "phi3-onnx-chat-cs", "flag1": "{_IS_PHI3_ONNX_CHAT_STREAMING_TEMPLATE}", "flag2": "{_IS_PHI3_ONNX_CHAT_FUNCTIONS_TEMPLATE}" }
+
     # Speech to Text
     # - { "dir1": "speech-to-text-continuous-reco-cs", "dir2": "speech-to-text-cs",                  "file1": "SpeechToTextContinuousReco.csproj._", "file2": "SpeechToText.csprj._",                "file3": "SpeechToText.csproj._",                "dir3": "speech-to-text-cs", "flag1": "{_IS_SPEECH_TO_TEXT_CONTINUOUS}", "flag2": "!{_IS_SPEECH_TO_TEXT_CONTINUOUS}" }
     # - { "dir1": "speech-to-text-continuous-reco-cs", "dir2": "speech-to-text-cs",                  "file1": "Program.cs",                          "file2": "Program.cs",                          "file3": "Program.cs",                           "dir3": "speech-to-text-cs", "flag1": "{_IS_SPEECH_TO_TEXT_CONTINUOUS}", "flag2": "!{_IS_SPEECH_TO_TEXT_CONTINUOUS}" }
@@ -27,8 +33,8 @@
     # - { "dir1": "speech-to-text-continuous-reco-cs", "dir2": "speech-to-text-with-translation-cs", "file1": "Program.cs",                          "file2": "Program.cs",                          "file3": "Program.cs",                           "dir3": "speech-to-text-cs", "flag1": "{_IS_SPEECH_TO_TEXT_CONTINUOUS}", "flag2": "{_IS_SPEECH_TO_TEXT_WITH_TRANSLATION}" }
 
     # Text to Speech
-    - { "dir1": "text-to-speech-cs", "dir2": "text-to-speech-with-file-cs", "file1": "TextToSpeech.csproj._", "file2": "TextToSpeechWithFile.csproj._", "file3": "TextToSpeech.csproj._", "dir3": "text-to-speech-cs", "flag1": "!{_IS_TEXT_TO_SPEECH_WITH_FILE_TEMPLATE}", "flag2": "{_IS_TEXT_TO_SPEECH_WITH_FILE_TEMPLATE}" }
-    - { "dir1": "text-to-speech-cs", "dir2": "text-to-speech-with-file-cs", "file1": "Program.cs",            "file2": "Program.cs",                    "file3": "Program.cs",            "dir3": "text-to-speech-cs", "flag1": "!{_IS_TEXT_TO_SPEECH_WITH_FILE_TEMPLATE}", "flag2": "{_IS_TEXT_TO_SPEECH_WITH_FILE_TEMPLATE}" }
+    # - { "dir1": "text-to-speech-cs", "dir2": "text-to-speech-with-file-cs", "file1": "TextToSpeech.csproj._", "file2": "TextToSpeechWithFile.csproj._", "file3": "TextToSpeech.csproj._", "dir3": "text-to-speech-cs", "flag1": "!{_IS_TEXT_TO_SPEECH_WITH_FILE_TEMPLATE}", "flag2": "{_IS_TEXT_TO_SPEECH_WITH_FILE_TEMPLATE}" }
+    # - { "dir1": "text-to-speech-cs", "dir2": "text-to-speech-with-file-cs", "file1": "Program.cs",            "file2": "Program.cs",                    "file3": "Program.cs",            "dir3": "text-to-speech-cs", "flag1": "!{_IS_TEXT_TO_SPEECH_WITH_FILE_TEMPLATE}", "flag2": "{_IS_TEXT_TO_SPEECH_WITH_FILE_TEMPLATE}" }
 
   tests:
   - name: refactor
