+++ src/ai/.x/templates/includes/phi3-onnx-chat-cs/get-phi3-mini-onnx.cmd
@@ -1,0 +1,3 @@
+git lfs install
+git clone https://huggingface.co/microsoft/Phi-3-mini-4k-instruct-onnx
+git lfs checkout
