+++ src/ai/.x/templates/phi3-onnx-chat-streaming-cs/Phi3ChatStreaming.csproj._
@@ -1,17 +1,1 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-  <ItemGroup>
-    {{if contains(toupper("{ONNX_GENAI_MODEL_PLATFORM}"), "DIRECTML")}}
-    <PackageReference Include="Microsoft.ML.OnnxRuntimeGenAI.DirectML" Version="0.3.0" />
-    {{else if contains(toupper("{ONNX_GENAI_MODEL_PLATFORM}"), "CUDA")}}
-    <PackageReference Include="Microsoft.ML.OnnxRuntimeGenAI.Cuda" Version="0.3.0" />
-    {{else}}
-    <PackageReference Include="Microsoft.ML.OnnxRuntimeGenAI" Version="0.3.0" />
-    {{endif}}
-  </ItemGroup>
-</Project>
+{{@include phi3-onnx-chat-cs/Phi3ChatStreaming.csproj._}}
