+++ tests/test.yaml
@@ -103,8 +103,9 @@
       ^OpenAI +Chat +Webpage +\(w/ +Functions\) +openai-chat-webpage-with-functions +JavaScript, +TypeScript *\r?$\n
       ^OpenAI +Chat +Webpage +\(w/ +Speech +input/output\) +openai-chat-webpage-with-speech +TypeScript *\r?$\n
       ^OpenAI +Chat +Webpage +\(w/ +Functions +\+ +Speech\) +openai-chat-webpage-with-speech-and-functions +TypeScript *\r?$\n
-      ^OpenAI +Realtime +Chat +openai-realtime-chat +C# *\r?$\n
-      ^OpenAI +Realtime +Chat \(w/ Functions\) +openai-realtime-chat-with-functions +C# *\r?$\n
+      ^OpenAI Realtime Chat +openai-realtime-chat +C# *\r?$\n
+      ^OpenAI Realtime Chat \(w/ Functions\) +openai-realtime-chat-with-functions +C# *\r?$\n
+      ^OpenAI Realtime Chat \(w/ Keyword\) +openai-realtime-chat-with-keyword +C# *\r?$\n
       ^Phi-3 Chat Completions \(w/ ONNX\) +phi3-onnx-chat-streaming +C# *\r?$\n
       ^Phi-3 Chat Completions \(w/ ONNX \+ Functions\) +phi3-onnx-chat-streaming-with-functions +C# *\r?$\n
       ^Semantic +Kernel +Chat +Completions +\(Streaming\) +sk-chat-streaming +C# *\r?$\n
