+++ tests/test3.yaml
@@ -1684,3 +1684,27 @@
         Tell me a joke
         exit
       tag: skip
+
+- area: ai dev new phi3-onnx-chat-streaming-with-functions (c#)
+  tests:
+
+  - class: dev new phi3-onnx-chat-streaming-with-functions (c#)
+    steps:
+    - name: generate template
+      command: ai dev new phi3-onnx-chat-streaming-with-functions --cs
+    - name: build template
+      bash: |
+        cd phi3-onnx-chat-streaming-with-functions-cs
+        dotnet build
+    - name: get models
+      script: |
+        cd phi3-onnx-chat-streaming-with-functions-cs
+        ./get-phi3-mini-onnx.cmd
+      tag: skip
+    - name: run template
+      command: ai dev shell --bash "cd phi3-onnx-chat-streaming-with-functions-cs;./bin/Debug/net8.0/Phi3ChatStreaming"
+      input: |-
+        What is the date?
+        What is the time?
+        exit
+      tag: skip
